Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-    int a, b, c;
+    float a, b, c;
     float avg;
-    scanf("%d %d %d %f", &a, &b, &c, &avg);
+    scanf("%f %f %f %f", &a, &b, &c, &avg);
     avg = (a+b+c)/3;
     printf("Average: %.2f\n", avg);
     return 0;
